requirements.txt
# Task 1 – Telegram Scraping
telethon>=1.38
python-dotenv>=1.0
loguru>=0.7

# Task 2 – Data Loading & dbt
pandas>=2.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
dbt-postgres>=1.6

# Task 4 – FastAPI (API Layer)
fastapi>=0.110
uvicorn>=0.27
pydantic>=2.6

# Task 5 – Orchestration
dagster>=1.6
dagster-webserver>=1.6

# Task 3 – Image Enrichment
ultralytics>=8.0

# Utilities / Testing
httpx>=0.27
